// CPQ12/src/types/pricing.ts
export interface UserLimits {
  from: number;
  to: number;
}

export interface GBLimits {
  from: number;
  to: number;
}

export interface PricingTier {
  id: string;
  name: 'Basic' | 'Standard' | 'Advanced';
  perUserCost: number;
  perGBCost: number;
  managedMigrationCost: number;
  instanceCost: number;
  userLimits: UserLimits;
  gbLimits: GBLimits;
  features: string[];
}